requirements.txt
pandas
numpy
langchain
python-dotenv
gradio
tqdm
openai
chromadb
tiktoken